Adding "extras" field to audio and photo tables.
