+++ apps/site/api/serializers/marker_serializer.py
@@ -5,22 +5,8 @@ from localground.apps.site import models, widgets
 from localground.apps.site.api import fields
 from django.conf import settings
 from localground.apps.site.api.metadata import CustomMetadata
-from localground.apps.site.api.fields.json_fields import JSONField
 
-class MarkerSerializerMixin(GeometrySerializer):
-    geometry = fields.GeometryField(
-        help_text='Assign a GeoJSON string',
-        allow_null=True,
-        source='point',
-        required=False,
-        style={'base_template': 'json.html'})
-    
-    extras = JSONField(
-        help_text='Store arbitrary key / value pairs here in JSON form. Example: {"key": "value"}',
-        allow_null=True,
-        required=False,
-        style={'base_template': 'json.html'})
-    
+class MarkerSerializerMixin(GeometrySerializer):    
     color = fields.ColorField(required=False)
     update_metadata = serializers.SerializerMethodField()
