+++ apps/site/models/video.py
@@ -1,10 +1,10 @@
 from django.contrib.gis.db import models
 from localground.apps.site.managers import VideoManager
-from localground.apps.site.models import BasePoint, BaseUploadedMedia
+from localground.apps.site.models import ExtrasMixin, BasePointMixin, BaseUploadedMedia
 import os
 
 
-class Video(BasePoint, BaseUploadedMedia):
+class Video(ExtrasMixin, BasePointMixin, BaseUploadedMedia):
     name = 'video'
     name_plural = 'videos'
     objects = VideoManager()
